2 - Add functionality to main programm
