+++ src/main/java/com/example/insurance/services/InsuranceService.java
@@ -1,6 +1,7 @@
 package com.example.insurance.services;
 
 import com.example.insurance.model.VehicleCalcResult;
+import com.example.insurance.services.calculation.strategy.CalculationStrategy;
 
 import java.util.List;
