+++ src/main/java/com/example/insurance/services/calculation/strategy/CalculationStrategyByCarProducerCoeffVehicleAgeAndVehicleValue.java
@@ -1,4 +1,4 @@
-package com.example.insurance.services;
+package com.example.insurance.services.calculation.strategy;
 
 import com.example.insurance.model.Vehicle;
 import com.example.insurance.repositories.CoefficientRepository;
