+++ src/test/java/com/example/insurance/services/InsuranceServiceImplTest.java
@@ -3,6 +3,7 @@ package com.example.insurance.services;
 import com.example.insurance.model.Vehicle;
 import com.example.insurance.model.VehicleCalcResult;
 import com.example.insurance.repositories.VehicleRepository;
+import com.example.insurance.services.calculation.strategy.CalculationStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -17,11 +18,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class InsuranceServiceImplTest {
-
-    private static final double TEST_CAR_PRODUCER_RISK_COEFF = 1.3;
-    private static final double TEST_VEHICLE_AGE_RISK_COEFF = 1.1;
-    private static final double TEST_VEHICLE_VALUE_RISK_COEFF = 0.08;
-    private static final double TEST_VEHICLE_PREVIOUS_INDEMNITY_RISK_COEFF = 0.05;
 
     private static final String TEST_VEHICLE_PLATE_NUMBER = "AAA000";
     private static final int TEST_VEHICLE_REGISTRATION = 2019;
