+++ src/test/java/com/example/insurance/services/calculation/strategy/CalculationStrategyByCarProducerCoeffVehicleAgeAndVehicleValueTest.java
@@ -1,4 +1,4 @@
-package com.example.insurance.services;
+package com.example.insurance.services.calculation.strategy;
 
 import com.example.insurance.model.Vehicle;
 import com.example.insurance.model.VehicleCalcResult;
@@ -10,10 +10,10 @@ import java.time.LocalDate;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-class CalculationStrategyByCarProducerCoeffVehicleAgeVehicleValueAndPreviousIndemnityTest extends CalculationStrategyBaseTest {
+class CalculationStrategyByCarProducerCoeffVehicleAgeAndVehicleValueTest extends CalculationStrategyBaseTest {
 
     @InjectMocks
-    CalculationStrategyByCarProducerCoeffVehicleAgeVehicleValueAndPreviousIndemnity calculationStrategy;
+    CalculationStrategyByCarProducerCoeffVehicleAgeAndVehicleValue calculationStrategy;
 
     @Test
     void getVehicleCalcResult() {
@@ -21,7 +21,7 @@ class CalculationStrategyByCarProducerCoeffVehicleAgeVehicleValueAndPreviousInde
 
         checkVehicleCalcResult(result);
 
-        double expectedAnnualFee = TEST_CAR_PRODUCER_RISK_COEFF * (TEST_VEHICLE_AGE_RISK_COEFF * (LocalDate.now().getYear() - TEST_VEHICLE_REGISTRATION) + TEST_VEHICLE_VALUE_RISK_COEFF * TEST_VEHICLE_PURCHASE_PRICE + TEST_VEHICLE_PREVIOUS_INDEMNITY * TEST_VEHICLE_PREVIOUS_INDEMNITY_RISK_COEFF);
+        double expectedAnnualFee = TEST_CAR_PRODUCER_RISK_COEFF * (TEST_VEHICLE_AGE_RISK_COEFF * (LocalDate.now().getYear() - TEST_VEHICLE_REGISTRATION) + TEST_VEHICLE_VALUE_RISK_COEFF * TEST_VEHICLE_PURCHASE_PRICE);
 
         assertEquals(expectedAnnualFee, result.getAnnualFee());
         assertEquals(expectedAnnualFee / 12, result.getMonthlyFee());
@@ -35,4 +35,5 @@ class CalculationStrategyByCarProducerCoeffVehicleAgeVehicleValueAndPreviousInde
 
         assertNull(result);
     }
+
 }
